ll_schedule: refine the _domain_set/clear() logic
Refine and simplify the domain_set/clear() part:

1. Use UINT64_MAX to denote that the domain is not set.

2. For _domain_set(), set only when needed. e.g. the first task over all
cores being added, or a task asking for earlier timer being added.

The _domain_set() logic is refined as below:
a. For first task, set domain according to the required start, else
b. for one shot task, set domain if it is earlier, else
c. for earlier periodic task, try to make it cadence-aligned with the
existed task, else
d. for later periodic task, simply cover it by the coming interrupt.

3. For _domain_clear() logic, simplify to call it when the last client is
freed.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>